+++ internal/selfupdate/replace_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package selfupdate
+
+import "os"
+
+func replaceExecutable(stagedPath, executablePath string) error {
+	return os.Rename(stagedPath, executablePath)
+}